requirements.txt
tensorflow==2.9.2
numpy
pandas
transformers
datasets
rouge-score
nltk
huggingface_hub
jupyter
from-root
boto3
mypy-boto3-s3
fastapi
uvicorn
Jinja2
Pillow
-e .